get rid of the overlap in the bbox.  add with clause option
I can't see that the overlap thing makes any sense at all, so I am
getting rid of it.  perhaps put it back if there is a problem with tiling

Added get_bbox_with_format.  What this function does is essentially the same
as get_bbox, but it will make it so that you can easily insert it in a with clause,
like so

var with_clause = 'with ' + geom_utils.get_bbox_with_format(req,'bbox_area')

where the second argument is the alias to use in the sql query for the with clause.